+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-hangelog
+Changelog
 =========
 
-* **2013-05-30**: [DependencyInjection] added `enabled` flag to `phpcr_provider` config
+* **2013-07-28**: [DependencyInjection] added `enabled` flag to `phpcr_provider` config
 * **2013-07-26**: [Model] Deprecated setRouteContent and getRouteContent, use setContent and getContent instead.
 * **2013-07-19**: [Model] Separated database agnostic, doctrine generic and
   PHPCR-ODM specific code to prepare for Doctrine ORM support.