+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+1.0.0-RC1
+---------
+
 * **2013-07-31**: [EventDispatcher] Added events to the dynamic router at the start of match and matchRequest
 * **2013-07-29**: [DependencyInjection] restructured `phpcr_provider` config into `persistence` -> `phpcr` to match other Bundles
 * **2013-07-28**: [DependencyInjection] added `enabled` flag to `phpcr_provider` config
@@ -10,6 +13,7 @@ Changelog
 * **2013-07-17**: [FormType] Moved TermsFormType to CoreBundle and renamed it to CheckboxUrlLableFormType
 
 1.1.0-beta2
+-----------
 
 * **2013-05-28**: [Bundle] Only include Doctrine PHPCR compiler pass if PHPCR-ODM is present
 * **2013-05-25**: [Bundle] Drop symfony_ from symfony_cmf prefix